project: remove lbcli-rs (#12)
The official lockbook CLI was heavily refactored in #1561 to include much of the improved functionality and cleaner code in this one, such as much simpler error handling that doesn't get in the way, passing either a UUID or path as positional arguments, better output for ls and other commands, and much more usable share functionalities. Therefore, lbcli-rs has served its purpose.